gallery added to menu
